+++ CHANGES.md
@@ -21,6 +21,9 @@
 - [UPDATE] AudioDeviceBuffer の変更に追随し、初期化時に env_ を渡すよう修正する
   - libwebrtc アップデートによって AudioDeviceBuffer が env を直接参照するようになったため
   - @torikizi
+- [UPDATE] boost::json のリンクエラーを修正する
+  - util.cpp の CodecCapability 出力で boost::json::string をそのままストリーム出力しないよう `value_to<std::string>` へ変更
+  - @torikizi
 
 ## 2025.3.0